# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/analytics/overview": {
      "get": {
        "operationId": "AdminAnalyticsController_getFinancialOverview",
        "parameters": [
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Financial overview retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get financial overview",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/investments": {
      "get": {
        "operationId": "AdminAnalyticsController_getInvestmentAnalytics",
        "parameters": [
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Investment analytics retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get investment analytics",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/profits": {
      "get": {
        "operationId": "AdminAnalyticsController_getProfitAnalytics",
        "parameters": [
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Profit analytics retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get profit analytics",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/revenue": {
      "get": {
        "operationId": "AdminAnalyticsController_getRevenueAnalytics",
        "parameters": [
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue analytics retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get revenue analytics",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/user-behavior": {
      "get": {
        "operationId": "AdminAnalyticsController_getUserBehaviorAnalytics",
        "parameters": [
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "User behavior analytics retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user behavior analytics",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/plan-performance": {
      "get": {
        "operationId": "AdminAnalyticsController_getPlanPerformance",
        "parameters": [
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Plan performance analytics retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get investment plan performance comparison",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/cash-flow": {
      "get": {
        "operationId": "AdminAnalyticsController_getCashFlowProjections",
        "parameters": [
          {
            "name": "days",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cash flow projections retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get cash flow projections",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/risk-exposure": {
      "get": {
        "operationId": "AdminAnalyticsController_getRiskExposure",
        "parameters": [
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Risk exposure metrics retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get risk exposure metrics",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/daily-report/{date}": {
      "get": {
        "operationId": "AdminAnalyticsController_getDailyReport",
        "parameters": [
          {
            "name": "date",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily report retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get daily financial report",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/daily-reports": {
      "get": {
        "operationId": "AdminAnalyticsController_getDailyReports",
        "parameters": [
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Daily reports retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get daily reports for date range",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/weekly-summary": {
      "get": {
        "operationId": "AdminAnalyticsController_getWeeklySummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Weekly summary retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get weekly summary",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/reports/generate": {
      "post": {
        "operationId": "AdminAnalyticsController_generateReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportGenerationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report generated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate custom report",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/reports/export-csv": {
      "post": {
        "operationId": "AdminAnalyticsController_exportCSV",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsQueryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data exported successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export data as CSV",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
    "/api/admin/analytics/daily-report/regenerate/{date}": {
      "post": {
        "operationId": "AdminAnalyticsController_regenerateReport",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily report regenerated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Regenerate daily report for specific date",
        "tags": [
          "Admin Analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "CvestAPI",
    "description": "API documentation for CvestAPI",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ReportGenerationDto": {
        "type": "object",
        "properties": {}
      },
      "AnalyticsQueryDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}